pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-gbill_env"
version = "0.1.0"
description = "Invoice-processing environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.3",
    "orjson>=3.11.7",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.2", "pytest-cov>=7.1.0"]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m gbill_env.server.app
server = "gbill_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["gbill_env", "gbill_env.server"]
package-dir = { "gbill_env" = ".", "gbill_env.server" = "server" }
